requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
simplejson==3.19.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
